Fix ANSI issues for Linux Ubunto doesn't like the BOLD_OFF (21) ANSI escape sequence. Rather than turning bold off, it switches the state to put a double-underline under EVERYTHING. Who thought this was a good idea, I've no clue, but it's an annoying divergence from the specified behavior that I had to work around.
I need to do a pass over the entire .bashrc_xtras. I feel like there's stuff in
there that isn't working as it should anymore. The Ubuntu terminal doesn't
support the status bar text anymore (which is another annoyance, given that it
was for a Ubuntu box that I first added all of this.) Also, I still have ANKI
stuff in there.